菲律宾单国信息简报/requirements-lock.txt
# 锁定依赖版本 - 用于服务器部署
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
urllib3==2.6.3
